+++ src/Models/Category/Category.php
@@ -35,6 +35,43 @@ class Category extends Model
         return $this->hasOne(CategoryData::class, 'cat_id', 'cat_id')->customFields($fields);
     }
 
+    public function scopeOrderByCustomField($query, $field, $direction = 'asc')
+    {
+        $manager = app(FieldtypeManager::class);
+
+        if (! $manager->hasField($field, 'category')) {
+            return $query;
+        }
+
+        $field = $manager->getField($field, 'category');
+        $column = "field_id_{$field->field_id}";
+
+        // If this field is not storing it's data on the category_field_data table we
+        // will join the separate data table with a unique orderby_field_name alias
+        if ($field->legacy_field_data == 'n' || $field->legacy_field_data === false) {
+            $alias = "orderby_{$field->field_name}";
+            $column = "$alias.$column";
+            $this->scopeJoinFieldDataTable($query, $field, $alias);
+        }
+
+        return $query->orderBy($column, $direction);
+    }
+
+    public function scopeJoinFieldDataTable($query, $field, $alias = null)
+    {
+        if ($field->legacy_field_data == 'y' || $field->legacy_field_data === true) {
+            return $query;
+        }
+
+        $table = $field->data_table_name;
+        $joinTable = $alias ? "$table as $alias" : $table;
+        $alias = $alias ?: $table;
+        $query->leftJoin($joinTable, "$alias.cat_id", '=', $this->qualifyColumn('cat_id'));
+        $query->select('*', $this->qualifyColumn('cat_id'));
+
+        return $query;
+    }
+
     public function files()
     {
         return $this->belongsToMany(File::class, 'file_categories', 'file_id', 'cat_id');
@@ -64,7 +101,7 @@ class Category extends Model
         // we will forward the check to the entry's data
         if (is_null($value) && app(FieldtypeManager::class)->hasField($key, 'category')) {
             $this->getRelationValue('data');
-            $fields = $this->data->fields($this->group);
+            $fields = $this->data->fields($this);
             $value = ($fields->has($key)) ? $fields->get($key) : new FieldContent([
                 // 'field' => $fields->get($key),
                 'field' => app(FieldtypeManager::class)->getField($key),